share one factory for the default model map creation
